requirements.txt
flask
flask-sqlalchemy
flask-login
flask-migrate
python-dotenv
mysql-connector-python
pymysql
cryptography
flask-mail
flask-babel
flask-wtf
flask-caching
